Use lazy handler for some lifecycle handlers at the top level driver
