+++ drivers/SmartThings/zigbee-switch/src/init.lua
@@ -20,10 +20,17 @@ local configurationMap = require "configurations"
 local CONFIGURE_REPORTING_RESPONSE_ID = 0x07
 local SIMPLE_METERING_ID = 0x0702
 local ELECTRICAL_MEASUREMENT_ID = 0x0B04
+local version = require "version"
+
+local lazy_handler
+if version.api >= 15 then
+  lazy_handler = require "st.utils.lazy_handler"
+else
+  lazy_handler = require
+end
 
 local function lazy_load_if_possible(sub_driver_name)
   -- gets the current lua libs api version
-  local version = require "version"
 
   -- version 9 will include the lazy loading functions
   if version.api >= 9 then
@@ -32,24 +39,6 @@ local function lazy_load_if_possible(sub_driver_name)
     return require(sub_driver_name)
   end
 
-end
-
-local function info_changed(self, device, event, args)
-  local preferences = require "preferences"
-  preferences.update_preferences(self, device, args)
-end
-
-local do_configure = function(self, device)
-  device:refresh()
-  device:configure()
-
-  -- Additional one time configuration
-  if device:supports_capability(capabilities.energyMeter) or device:supports_capability(capabilities.powerMeter) then
-    local clusters = require "st.zigbee.zcl.clusters"
-    -- Divisor and multipler for EnergyMeter
-    device:send(clusters.SimpleMetering.attributes.Divisor:read(device))
-    device:send(clusters.SimpleMetering.attributes.Multiplier:read(device))
-  end
 end
 
 local function component_to_endpoint(device, component_id)
@@ -64,10 +53,6 @@ local function endpoint_to_component(device, ep)
   else
     return "main"
   end
-end
-
-local function find_child(parent, ep_id)
-  return parent:get_child_by_parent_assigned_key(string.format("%02X", ep_id))
 end
 
 local device_init = function(driver, device)
@@ -87,51 +72,10 @@ local device_init = function(driver, device)
   end
   local device_lib = require "st.device"
   if device.network_type == device_lib.NETWORK_TYPE_ZIGBEE then
+    local find_child = require "lifecycle_handlers.find_child"
     device:set_find_child(find_child)
   end
 end
-
-local function is_mcd_device(device)
-  local components = device.profile.components
-  if type(components) == "table" then
-    local component_count = 0
-    for _, component in pairs(components) do
-        component_count = component_count + 1
-    end
-    return component_count >= 2
-  end
-end
-
-local function device_added(driver, device, event)
-  local clusters = require "st.zigbee.zcl.clusters"
-  local device_lib = require "st.device"
-
-  local main_endpoint = device:get_endpoint(clusters.OnOff.ID)
-  if is_mcd_device(device) == false and device.network_type == device_lib.NETWORK_TYPE_ZIGBEE then
-    for _, ep in ipairs(device.zigbee_endpoints) do
-      if ep.id ~= main_endpoint then
-        if device:supports_server_cluster(clusters.OnOff.ID, ep.id) then
-          device:set_find_child(find_child)
-          if find_child(device, ep.id) == nil then
-            local name = string.format("%s %d", device.label, ep.id)
-            local child_profile = "basic-switch"
-            driver:try_create_device(
-              {
-                type = "EDGE_CHILD",
-                label = name,
-                profile = child_profile,
-                parent_device_id = device.id,
-                parent_assigned_child_key = string.format("%02X", ep.id),
-                vendor_provided_label = name
-              }
-            )
-          end
-        end
-      end
-    end
-  end
-end
-
 
 local zigbee_switch_driver_template = {
   supported_capabilities = {
@@ -186,9 +130,9 @@ local zigbee_switch_driver_template = {
   current_config_version = 1,
   lifecycle_handlers = {
     init = configurationMap.power_reconfig_wrapper(device_init),
-    added = device_added,
-    infoChanged = info_changed,
-    doConfigure = do_configure
+    added = lazy_handler("lifecycle_handlers.device_added"),
+    infoChanged = lazy_handler("lifecycle_handlers.info_changed"),
+    doConfigure = lazy_handler("lifecycle_handlers.do_configure"),
   },
   health_check = false,
 }
